fix: Resolve env file in runtime
