+++ index.ts
@@ -1,1 +1,2 @@
 export * from './src/Env'
+export * from './src/utils/resolveEnvFile'
